api/requirements.txt
google-genai
fastapi
pydantic
mangum
